Restructure roles components and containers into plugin modules.
