+++ src/app/components/SheriffDutyBar/SheriffDutyBar.css
@@ -1,5 +1,5 @@
 .sheriff-duty-bar{
-    height: 100%;  
+    height: 100%;
     display: flex;
     align-content: stretch;
     align-items: stretch;
@@ -28,7 +28,7 @@
         background-color: rgba(255, 255, 255, .2);
 }
 
-/* Don't Display the overlay if we click on the duty-bar */
+/* Don't TableColumnCellDisplay the overlay if we click on the duty-bar */
 .sheriff-duty-bar:active::before{
     display: none;
 }
