+++ src/app/components/SheriffUnassignedBar/SheriffUnassignedBar.css
@@ -1,5 +1,5 @@
 .sheriff-unassigned-bar{
-    height: 50%;  
+    height: 50%;
     display: flex;
     align-content: stretch;
     align-items: stretch;
@@ -29,7 +29,7 @@
         background-color: rgba(255, 255, 255, .2);
 } */
 
-/* Don't Display the overlay if we click on the duty-bar */
+/* Don't TableColumnCellDisplay the overlay if we click on the duty-bar */
 .sheriff-unassigned-bar:active::before{
     display: none;
 }
