+++ src/app/plugins/AdminRoles/AdminRoles.tsx
@@ -67,17 +67,17 @@ import {
 } from '../../components/Form/FormContainer';
 
 import DataTable, { DetailComponentProps, EmptyDetailRow } from '../../components/Table/DataTable';
-import AdminRoleScopeAccessModal from '../../containers/AdminRolesGrid/AdminRoleScopeAccessModal';
+import AdminRoleScopeAccessModal from './components/AdminRoleScopeAccessModal';
 
 // import RoleSelector from './RoleSelector';
 // import FrontendScopeDisplay from './FrontendScopeDisplay';
-import FrontendScopeCodeDisplay from './FrontendScopeCodeDisplay';
-import FrontendScopeDescriptionDisplay from './FrontendScopeDescriptionDisplay';
+import FrontendScopeCodeDisplay from './containers/FrontendScopeCodeDisplay';
+import FrontendScopeDescriptionDisplay from './containers/FrontendScopeDescriptionDisplay';
 // import ApiScopeDisplay from './ApiScopeDisplay';
-import ApiScopeCodeDisplay from './ApiScopeCodeDisplay';
-import ApiScopeDescriptionDisplay from './ApiScopeDescriptionDisplay';
-import FrontendScopeSelector from './FrontendScopeSelector';
-import ApiScopeSelector from './ApiScopeSelector';
+import ApiScopeCodeDisplay from './containers/ApiScopeCodeDisplay';
+import ApiScopeDescriptionDisplay from './containers/ApiScopeDescriptionDisplay';
+import FrontendScopeSelector from './containers/FrontendScopeSelector';
+import ApiScopeSelector from './containers/ApiScopeSelector';
 import { RoleFrontendScopePermission } from '../../api/Api';
 import RemoveRow from '../../components/TableColumnActions/RemoveRow';
 import ExpireRow from '../../components/TableColumnActions/ExpireRow';
@@ -150,8 +150,13 @@ class RoleApiScopesDataTable extends DataTable<RoleApiScope> {}
 
 let RENDER_COUNT = 0;
 
-export default class AdminRolesGrid extends FormContainerBase<AdminRolesProps> {
-    name = 'admin-roles-grid';
+export default class AdminRoles extends FormContainerBase<AdminRolesProps> {
+    // NOTICE!
+    // This key maps to the [appScope: FrontendScope] (in the token)
+    // To set permissions for a new plugin, add a corresponding entry under System Settings > Components
+    // with the name as defined as the plugin's name.
+    name = 'ADMIN_PLUGIN_ROLES';
+    // END NOTICE
     reduxFormKey = 'roles';
     formFieldNames = {
         roles: 'roles.roles',
