+++ src/app/plugins/AdminRoles/components/AdminScopePermissionsModal.tsx
@@ -4,13 +4,13 @@ import {
     Button,
     Glyphicon
 } from 'react-bootstrap';
-import ModalWrapper from '../ModalWrapper/ModalWrapper';
-import DataTable, { EmptyDetailRow } from '../../components/Table/DataTable';
+import ModalWrapper from '../../../containers/ModalWrapper/ModalWrapper';
+import DataTable, { EmptyDetailRow } from '../../../components/Table/DataTable';
 
-import SelectorField from '../../components/FormElements/SelectorField';
-import FrontendScopeSelector from './FrontendScopeSelector';
-import RemoveRow from '../../components/TableColumnActions/RemoveRow';
-import ExpireRow from '../../components/TableColumnActions/ExpireRow';
+import SelectorField from '../../../components/FormElements/SelectorField';
+import FrontendScopeSelector from '../containers/FrontendScopeSelector';
+import RemoveRow from '../../../components/TableColumnActions/RemoveRow';
+import ExpireRow from '../../../components/TableColumnActions/ExpireRow';
 
 export interface AdminScopePermissionsModalProps {
     isOpen?: boolean;
