+++ src/app/plugins/AdminRoles/containers/ApiScopeDescriptionDisplay.tsx
@@ -1,20 +1,20 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { SelectorProps } from '../../components/FormElements/Selector';
+import { SelectorProps } from '../../../components/FormElements/Selector';
 
-import { ApiScope } from '../../api';
+import { ApiScope } from '../../../api';
 
-import { RootState } from '../../store';
-import { getAllApiScopes } from '../../modules/roles/selectors';
+import { RootState } from '../../../store';
+import { getAllApiScopes } from '../../../modules/roles/selectors';
 
-interface ApiScopeCodeDisplayStateProps {
+interface ApiScopeDescriptionDisplayStateProps {
     apiScopes?: ApiScope[];
     input?: any;
 }
 
-class ApiScopeCodeDisplay extends React.PureComponent<
-    ApiScopeCodeDisplayStateProps & SelectorProps> {
+class ApiScopeDescriptionDisplay extends React.PureComponent<
+    ApiScopeDescriptionDisplayStateProps & SelectorProps> {
 
     render() {
         const {
@@ -27,7 +27,7 @@ class ApiScopeCodeDisplay extends React.PureComponent<
         if (value) {
             const values = apiScopes.map(scope => ({
                 key: scope.id as string,
-                value: scope.scopeCode as string
+                value: scope.description as string
             }));
 
             const match = values.find((item) => item.key === value);
@@ -50,6 +50,6 @@ const mapStateToProps = (state: RootState) => {
 };
 
 // tslint:disable-next-line:max-line-length
-export default connect<ApiScopeCodeDisplayStateProps, {}, SelectorProps>(
+export default connect<ApiScopeDescriptionDisplayStateProps, {}, SelectorProps>(
     mapStateToProps
-)(ApiScopeCodeDisplay);
+)(ApiScopeDescriptionDisplay);
