+++ src/app/plugins/AdminRoles/containers/RoleDisplay.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import Selector, { SelectorProps } from '../../components/FormElements/Selector';
+import Selector, { SelectorProps } from '../../../components/FormElements/Selector';
 
-import { Role } from '../../api';
+import { Role } from '../../../api';
 
-import { RootState } from '../../store';
-import { getAllRoles } from '../../modules/roles/selectors';
+import { RootState } from '../../../store';
+import { getAllRoles } from '../../../modules/roles/selectors';
 
 interface RoleDisplayStateProps {
     roles?: Role[];
